Update Ashu_Talif_Captain.lua
Added an onMobEngaged function and copied the existing logic that was previously placed under onMobRoam under there. The mob would never roam as he would either spawn engaged to the player, or if the player is out of range, spawn engaged to Gessho. I do not feel as though this is the ideal solution but this will significantly reduce the number of complaints of the fight not being able to be beaten.